communication: HMAC for BitBox >= v5.0.0
Since BitBox v5.0.0, messages send to and received from the BitBox
are authenticated. Therefore, the application appends an HMAC
prior to sending a command to the BitBox and verifies a received HMAC
prior to decrypting the data received from the BitBox.
The fix is backwards-compatible so that BitBoxes <v5.0.0 continue to work.